requirements.txt
langchain
openai
ipykernel
pandas
tabulate
gradio
python-dotenv
